Start drawing circular arcs segments of circles.
Wrote a structure for SVGArc based on the SVG path specification for
arcs
(https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths#arcs).
Wrote a function that prints the contents of SVGArc in a line that can
be inserted into an SVG file.

Wrote a function that takes a circle (structure with xy coordinates and
        radius) plus a half-cone angle and an incidence angle, and
returns SVGARcs. Did some basic manual tests and it seems to be behaving
how I'd like it to.