finish parse_viewbox_extents
Parse viewbox extents changed to always return a tuple of four values
(xmin, ymin, xmax, ymax). Added DEFAULT_WIDTH and DEFAULT_HEIGHT
constants to the top of the code, so that we have somewhere from which
to project a light source if needed. Currently handles svg files that
define any combination of "viewBox", "width", and "height".